updated file handling and added Exception Handling
previously to opening a file code was

  f = open("quotes.txt")
  quotes = f.readlines()
  f.close()

using this will work but not good practice. Use "context manager" to handle files in python and its easy.

changed variable name last to countNoQuotes. And it is not hardcoded anymore. Previously it was last = 13, now because f.readlines() return a list so if we aply len function to find length of the quotes list then len(quotes) will be count of all quotes.

Using print(quotes[rnd], end='') it will not print newline ("\n")